Move implement to gameboy.c
